+++ __tests__/cases/server/server.js
@@ -41,6 +41,7 @@ const CONFIG_FIELDS = [
 	'createMissingComponents',
 	'generatedComponentsExtension',
 	'sourceStylesDir',
+	'connectSocketAutomatically',
 ];
 
 describe('Server instance', () => {