+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Application has `log[Info|Warning|Error]` methods to prevent using console.
 - Server has `getConfig` method to property access to the server configuration.
 - Option to change location of source styles that are merged into the application styles.
+- Option to disable automatically socket connection.
 ### Updates
 - Creating subdirectories if needed while creating components.